+++ tests/test_filters.py
@@ -6,7 +6,7 @@ from __future__ import absolute_import, division, print_function
 
 import pytest
 
-from attr._make import attributes, attr
+from attr._make import attributes, attr, fields
 from attr.filters import _split_what, include, exclude
 
 
@@ -37,28 +37,28 @@ class TestInclude(object):
     @pytest.mark.parametrize("incl,value", [
         ((int,), 42),
         ((str,), "hello"),
-        ((str, C.a), 42),
-        ((str, C.b), "hello"),
+        ((str, fields(C).a), 42),
+        ((str, fields(C).b), "hello"),
     ])
     def test_allow(self, incl, value):
         """
         Return True if a class or attribute is whitelisted.
         """
         i = include(*incl)
-        assert i(C.a, value) is True
+        assert i(fields(C).a, value) is True
 
     @pytest.mark.parametrize("incl,value", [
         ((str,), 42),
         ((int,), "hello"),
-        ((str, C.b), 42),
-        ((int, C.b), "hello"),
+        ((str, fields(C).b), 42),
+        ((int, fields(C).b), "hello"),
     ])
     def test_drop_class(self, incl, value):
         """
         Return False on non-whitelisted classes and attributes.
         """
         i = include(*incl)
-        assert i(C.a, value) is False
+        assert i(fields(C).a, value) is False
 
 
 class TestExclude(object):
@@ -68,25 +68,25 @@ class TestExclude(object):
     @pytest.mark.parametrize("excl,value", [
         ((str,), 42),
         ((int,), "hello"),
-        ((str, C.b), 42),
-        ((int, C.b), "hello"),
+        ((str, fields(C).b), 42),
+        ((int, fields(C).b), "hello"),
     ])
     def test_allow(self, excl, value):
         """
         Return True if class or attribute is not blacklisted.
         """
         e = exclude(*excl)
-        assert e(C.a, value) is True
+        assert e(fields(C).a, value) is True
 
     @pytest.mark.parametrize("excl,value", [
         ((int,), 42),
         ((str,), "hello"),
-        ((str, C.a), 42),
-        ((str, C.b), "hello"),
+        ((str, fields(C).a), 42),
+        ((str, fields(C).b), "hello"),
     ])
     def test_drop_class(self, excl, value):
         """
         Return True on non-blacklisted classes and attributes.
         """
         e = exclude(*excl)
-        assert e(C.a, value) is False
+        assert e(fields(C).a, value) is False